# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Solver d'horaris Switch2",
...
    "contact": {
      "name": "HorarisSolver",
      "url": "https://github.com/XeviSanmartin/HorarisSolver"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/validate": {
      "post": {
        "tags": [
          "Validació"
        ],
        "summary": "Valida un Solver.json sense resoldre",
        "description": "Carrega les dades, executa les validacions i retorna advertiments i estadístiques.\n\nNo construeix ni resol el model: triga mil·lisegons i serveix per detectar\nproblemes abans de llançar una resolució. Els `advertiments` són informatius\ni **no** impedeixen cridar `/api/solve`.",
        "operationId": "validate_api_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticioValidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaValidate"
                }
              }
            }
          },
          "422": {
            "description": "El cos de la petició no compleix l'esquema (error Pydantic estàndard) o les dades tenen una incoherència estructural profunda (cos `DetallError`)."
          }
        }
      }
    },
    "/api/preprocess": {
      "post": {
        "tags": [
          "Validació"
        ],
        "summary": "Retorna les dades preprocessades (debug)",
        "description": "Retorna el JSON intermedi (`dades_solver_processades`) que consumeix el solver.\n\nÚtil per depurar què ha entès el preprocessador: mòduls especials detectats\n(FOL, anglès, sostenibilitat, digitalització), tutories, agrupacions\nfusionables i subgrups per curs.",
        "operationId": "preprocess_api_preprocess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticioValidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaPreprocess"
                }
              }
            }
          },
          "422": {
            "description": "El cos de la petició no compleix l'esquema (error Pydantic estàndard) o les dades tenen una incoherència estructural profunda (cos `DetallError`)."
          }
        }
      }
    },
    "/api/solve": {
      "post": {
        "tags": [
          "Resolució"
        ],
        "summary": "Resol l'horari",
        "description": "Executa el pipeline complet: preprocessament → model CP-SAT → solució.\n\nInterpreteu el camp `estat` de la resposta (vegeu la descripció general de\nl'API). Amb dades reals d'un centre, un `max_time_seconds` de 60–120 s sol\ndonar una solució `FEASIBLE`; si rebeu `UNKNOWN`, reintenteu amb més temps.",
        "operationId": "solve_api_solve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticioSolve"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaSolve"
                }
              }
            }
          },
          "422": {
            "description": "El cos de la petició no compleix l'esquema (error Pydantic estàndard) o les dades tenen una incoherència estructural profunda (cos `DetallError`)."
          },
          "500": {
            "description": "Error intern construint o resolent el model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetallError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DetallError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Descripció general del problema."
          },
          "detall": {
            "type": "string",
            "title": "Detall",
            "description": "Excepció concreta (tipus i missatge)."
          },
          "traca": {
            "type": "string",
            "title": "Traca",
            "description": "Traça reduïda per a depuració."
          }
        },
        "type": "object",
        "required": [
          "error",
          "detall",
          "traca"
        ],
        "title": "DetallError",
        "description": "Cos del camp `detail` en errors 422 (dades) i 500."
      },
...
      "OpcionsSolve": {
        "properties": {
          "max_time_seconds": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Time Seconds",
            "description": "Límit de temps de cerca en segons. El servidor l'acota al valor `max_temps_solver` que informa `GET /api/health`.",
            "default": 60
          },
          "num_workers": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Num Workers",
            "description": "Threads de cerca paral·lela de CP-SAT.",
            "default": 4
          },
          "incloure_compatible": {
            "type": "boolean",
            "title": "Incloure Compatible",
            "description": "Si és cert, la resposta inclou `solucio_compatible`: la solució en format Solver.json, reimportable a l'editor Switch2.",
            "default": false
          }
        },
        "type": "object",
        "title": "OpcionsSolve",
        "description": "Opcions d'execució del solver."
      },
      "PeticioSolve": {
        "properties": {
          "dades": {
            "$ref": "#/components/schemas/DadesSolver"
          },
          "opcions": {
            "$ref": "#/components/schemas/OpcionsSolve"
          }
        },
        "type": "object",
        "required": [
          "dades"
        ],
        "title": "PeticioSolve",
        "description": "Cos de la petició de /api/solve."
      },
      "PeticioValidate": {
        "properties": {
          "dades": {
            "$ref": "#/components/schemas/DadesSolver"
          }
        },
        "type": "object",
        "required": [
          "dades"
        ],
        "title": "PeticioValidate",
        "description": "Cos de la petició de /api/validate i /api/preprocess."
      },
...
      "RespostaPreprocess": {
        "properties": {
          "dades_processades": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dades Processades",
            "description": "JSON intermedi que consumeix el solver: professors amb restriccions normalitzades, mòduls amb marques `es_fol`/`es_angles`/…, cursos amb subgrups i tutor, agrupacions i configuració global."
          },
          "advertiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Advertiments"
          }
        },
        "type": "object",
        "required": [
          "dades_processades",
          "advertiments"
        ],
        "title": "RespostaPreprocess"
      },
      "RespostaSolve": {
        "properties": {
          "estat": {
            "type": "string",
            "title": "Estat",
            "description": "Resultat de la resolució: 'OPTIMAL', 'FEASIBLE', 'INFEASIBLE', 'UNKNOWN' o 'MODEL_INVALID'.",
            "examples": [
              "FEASIBLE"
            ]
          },
          "solucio": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Solucio"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present només si l'estat és OPTIMAL o FEASIBLE."
          },
          "solucio_compatible": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solucio Compatible",
            "description": "Solució en format Solver.json, reimportable a l'editor (present només si s'ha demanat amb `opcions.incloure_compatible`)."
          },
          "advertiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Advertiments",
            "description": "Advertiments del validador de dades (informatius)."
          }
        },
        "type": "object",
        "required": [
          "estat"
        ],
        "title": "RespostaSolve"
      },
      "RespostaValidate": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "description": "Cert si no hi ha cap advertiment."
          },
          "advertiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Advertiments",
            "description": "Incoherències detectades. NO bloquegen /api/solve.",
            "examples": [
              [
                "Mòdul Repàs no té cap professor assignat",
                "Cursos sense tutoria: DAMW D"
              ]
            ]
          },
          "estadistiques": {
            "additionalProperties": true,
            "type": "object",
            "title": "Estadistiques",
            "description": "Recompte d'entitats i mòduls especials detectats.",
            "examples": [
              {
                "moduls_angles": 4,
                "moduls_digitalizacio": 3,
                "moduls_fol": 5,
                "moduls_simultaneos": 2,
                "moduls_sostenibilitat": 3,
                "moduls_suport": 0,
                "subgrups_per_curs": {
                  "0": [
                    1,
                    2,
                    3
                  ]
                },
                "total_aules": 8,
                "total_cursos": 9,
                "total_moduls": 76,
                "total_professors": 18,
                "tutories": 9
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "valid",
          "advertiments",
          "estadistiques"
        ],
        "title": "RespostaValidate"
      },
...
      "Solucio": {
        "properties": {
          "horari": {
            "items": {},
            "type": "array",
            "title": "Horari",
            "description": "Vista per curs: `horari[curs][dia][hora]` → llista de classes (0 = lliure, 1 = classe, 2 = desdoblament de subgrups). Cada classe: `{modul, modul_index, professor, professor_index, aula, aula_index, subgrup}`."
          },
          "professors": {
            "items": {},
            "type": "array",
            "title": "Professors",
            "description": "Vista per professor: `professors[professor_index][dia][hora]` → llista de classes `{modul, modul_index, curs, curs_index, aula, aula_index, subgrup}`."
          },
          "aules": {
            "items": {},
            "type": "array",
            "title": "Aules",
            "description": "Vista per aula (indexada per posició, no per `aula_index`): `aules[pos][dia][hora]` → llista de classes `{modul, modul_index, professor, professor_index, curs, curs_index, subgrup}`."
          },
          "stats": {
            "$ref": "#/components/schemas/StatsSolucio"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "horari",
          "professors",
          "aules",
          "stats"
        ],
        "title": "Solucio",
        "description": "Horari resolt, en tres vistes indexades per [dia 0–4][hora 0–10]."
      },
...
    }
  },
...
}